Replace reboot_system class with cron in config.pp
The mymodule::reboot_system class has been removed and replaced with the mymodule::cron class in the puppet configuration file. This new class schedules a cron job for the 'baydev' user to execute the 'checkGit.sh' script upon reboot. This change aligns better with the need to frequently check for changes in the Git repositories.