Added cron module installation in puppet config file
A new class named 'install_cron_module' has been added to the puppet configuration file to install the 'cron_core' module using Puppet. This class ensures the module is installed only if it hasn't been done so already. This change caters to the need to ensure cron jobs function as expected.