Update file template path in Puppet configuration
The previous absolute path for file templates in the Puppet configuration file was replaced with a shorter one. This change streamlines the module installation commands, thus improving readability. It also enhances the adaptability of commands by relying on the environment's PATH, which eliminates potential issues when the binary location changes.